split context file into subfiles
