+++ kahypar/application/kahypar.cc
@@ -18,6 +18,7 @@
  *
  ******************************************************************************/
 
+#include "kahypar/partition/context/context_enum_classes.h"
 
 #include "kahypar/application/command_line_options.h"
 #include "kahypar/definitions.h"
