+++ kahypar/partition/context.h
@@ -30,7 +30,8 @@
 #include <vector>
 
 #include "kahypar/definitions.h"
-#include "kahypar/partition/context_enum_classes.h"
+#include "kahypar/partition/context/context_enum_classes.h"
+#include "kahypar/partition/context/include_parameters.h"
 #include "kahypar/partition/evolutionary/action.h"
 #include "kahypar/partition/parallel/communicator.h"
 #include "kahypar/utils/stats.h"
@@ -116,294 +117,6 @@ inline std::ostream& operator<< (std::ostream& str, const PreprocessingParameter
     str << params.community_detection;
   }
 
-  return str;
-}
-
-struct RatingParameters {
-  RatingFunction rating_function = RatingFunction::UNDEFINED;
-  CommunityPolicy community_policy = CommunityPolicy::UNDEFINED;
-  HeavyNodePenaltyPolicy heavy_node_penalty_policy = HeavyNodePenaltyPolicy::UNDEFINED;
-  AcceptancePolicy acceptance_policy = AcceptancePolicy::UNDEFINED;
-  RatingPartitionPolicy partition_policy = RatingPartitionPolicy::normal;
-  FixVertexContractionAcceptancePolicy fixed_vertex_acceptance_policy =
-    FixVertexContractionAcceptancePolicy::UNDEFINED;
-};
-
-inline std::ostream& operator<< (std::ostream& str, const RatingParameters& params) {
-  str << "  Rating Parameters:" << std::endl;
-  str << "    Rating Function:                  " << params.rating_function << std::endl;
-  str << "    Use Community Structure:          " << params.community_policy << std::endl;
-  str << "    Heavy Node Penalty:               " << params.heavy_node_penalty_policy << std::endl;
-  str << "    Acceptance Policy:                " << params.acceptance_policy << std::endl;
-  str << "    Partition Policy:                 " << params.partition_policy << std::endl;
-  str << "    Fixed Vertex Acceptance Policy:   " << params.fixed_vertex_acceptance_policy << std::endl;
-  return str;
-}
-
-
-struct CoarseningParameters {
-  CoarseningAlgorithm algorithm = CoarseningAlgorithm::UNDEFINED;
-  RatingParameters rating = { };
-  HypernodeID contraction_limit_multiplier = std::numeric_limits<HypernodeID>::max();
-  double max_allowed_weight_multiplier = std::numeric_limits<double>::max();
-
-  // Those will be determined dynamically
-  HypernodeWeight max_allowed_node_weight = 0;
-  HypernodeID contraction_limit = 0;
-  double hypernode_weight_fraction = 0.0;
-};
-
-inline std::ostream& operator<< (std::ostream& str, const CoarseningParameters& params) {
-  str << "Coarsening Parameters:" << std::endl;
-  str << "  Algorithm:                          " << params.algorithm << std::endl;
-  str << "  max-allowed-weight-multiplier:      " << params.max_allowed_weight_multiplier << std::endl;
-  str << "  contraction-limit-multiplier:       " << params.contraction_limit_multiplier << std::endl;
-  str << "  hypernode weight fraction:          ";
-  // For the coarsening algorithm of the initial partitioning phase
-  // these parameters are only known after main coarsening.
-  if (params.hypernode_weight_fraction == 0) {
-    str << "determined before IP";
-  } else {
-    str << params.hypernode_weight_fraction;
-  }
-  str << std::endl;
-  str << "  max. allowed hypernode weight:      ";
-  if (params.max_allowed_node_weight == 0) {
-    str << "determined before IP";
-  } else {
-    str << params.max_allowed_node_weight;
-  }
-  str << std::endl;
-  str << "  contraction limit:                  ";
-  if (params.contraction_limit == 0) {
-    str << "determined before IP";
-  } else {
-    str << params.contraction_limit;
-  }
-  str << std::endl << params.rating;
-  return str;
-}
-
-struct LocalSearchParameters {
-  struct FM {
-    uint32_t max_number_of_fruitless_moves = std::numeric_limits<uint32_t>::max();
-    double adaptive_stopping_alpha = std::numeric_limits<double>::max();
-    RefinementStoppingRule stopping_rule = RefinementStoppingRule::UNDEFINED;
-  };
-
-  struct Flow {
-    FlowAlgorithm algorithm = FlowAlgorithm::UNDEFINED;
-    FlowNetworkType network = FlowNetworkType::UNDEFINED;
-    FlowExecutionMode execution_policy = FlowExecutionMode::UNDEFINED;
-    double alpha = std::numeric_limits<double>::max();
-    size_t beta = std::numeric_limits<size_t>::max();
-    bool use_most_balanced_minimum_cut = false;
-    bool use_adaptive_alpha_stopping_rule = false;
-    bool ignore_small_hyperedge_cut = false;
-    bool use_improvement_history = false;
-  };
-
-  FM fm { };
-  Flow flow { };
-  RefinementAlgorithm algorithm = RefinementAlgorithm::UNDEFINED;
-  int iterations_per_level = std::numeric_limits<int>::max();
-};
-
-
-inline std::ostream& operator<< (std::ostream& str, const LocalSearchParameters& params) {
-  str << "Local Search Parameters:" << std::endl;
-  str << "  Algorithm:                          " << params.algorithm << std::endl;
-  str << "  iterations per level:               " << params.iterations_per_level << std::endl;
-  if (params.algorithm == RefinementAlgorithm::twoway_fm ||
-      params.algorithm == RefinementAlgorithm::kway_fm ||
-      params.algorithm == RefinementAlgorithm::kway_fm_km1 ||
-      params.algorithm == RefinementAlgorithm::twoway_fm_flow ||
-      params.algorithm == RefinementAlgorithm::kway_fm_flow_km1 ||
-      params.algorithm == RefinementAlgorithm::kway_fm_flow) {
-    str << "  stopping rule:                      " << params.fm.stopping_rule << std::endl;
-    if (params.fm.stopping_rule == RefinementStoppingRule::simple) {
-      str << "  max. # fruitless moves:             " << params.fm.max_number_of_fruitless_moves << std::endl;
-    } else {
-      str << "  adaptive stopping alpha:            " << params.fm.adaptive_stopping_alpha << std::endl;
-    }
-  }
-  if (params.algorithm == RefinementAlgorithm::twoway_flow ||
-      params.algorithm == RefinementAlgorithm::kway_flow ||
-      params.algorithm == RefinementAlgorithm::twoway_fm_flow ||
-      params.algorithm == RefinementAlgorithm::kway_fm_flow_km1 ||
-      params.algorithm == RefinementAlgorithm::kway_fm_flow) {
-    str << "  Flow Refinement Parameters:" << std::endl;
-    str << "    flow algorithm:                   " << params.flow.algorithm << std::endl;
-    str << "    flow network:                     " << params.flow.network << std::endl;
-    str << "    execution policy:                 " << params.flow.execution_policy << std::endl;
-    str << "    most balanced minimum cut:        "
-        << std::boolalpha << params.flow.use_most_balanced_minimum_cut << std::endl;
-    str << "    alpha:                            " << params.flow.alpha << std::endl;
-    if (params.flow.execution_policy == FlowExecutionMode::constant) {
-      str << "    beta:                             " << params.flow.beta << std::endl;
-    }
-    str << "    adaptive alpha stopping rule:     "
-        << std::boolalpha << params.flow.use_adaptive_alpha_stopping_rule << std::endl;
-    str << "    ignore small HE cut:              "
-        << std::boolalpha << params.flow.ignore_small_hyperedge_cut << std::endl;
-    str << "    use improvement history:          "
-        << std::boolalpha << params.flow.use_improvement_history << std::endl;
-  } else if (params.algorithm == RefinementAlgorithm::do_nothing) {
-    str << "  no coarsening!  " << std::endl;
-  }
-  return str;
-}
-
-struct InitialPartitioningParameters {
-  Mode mode = Mode::UNDEFINED;
-  InitialPartitioningTechnique technique = InitialPartitioningTechnique::UNDEFINED;
-  InitialPartitionerAlgorithm algo = InitialPartitionerAlgorithm::UNDEFINED;
-  CoarseningParameters coarsening = { };
-  LocalSearchParameters local_search = { };
-  uint32_t nruns = std::numeric_limits<uint32_t>::max();
-
-  // The following parameters are only used internally and are not supposed to
-  // be changed by the user.
-  PartitionID k = std::numeric_limits<PartitionID>::max();
-  HypernodeWeightVector upper_allowed_partition_weight = { };
-  HypernodeWeightVector perfect_balance_partition_weight = { };
-  PartitionID unassigned_part = 1;
-  // Is used to get a tighter balance constraint for initial partitioning.
-  // Before initial partitioning epsilon is set to init_alpha*epsilon.
-  double init_alpha = 1;
-  // If pool initial partitioner is used, the first 12 bits of this number decides
-  // which algorithms are used.
-  unsigned int pool_type = 1975;
-  // Maximum iterations of the Label Propagation IP over all hypernodes
-  int lp_max_iteration = 100;
-  // Amount of hypernodes which are assigned around each start vertex (LP)
-  int lp_assign_vertex_to_part = 5;
-  bool refinement = true;
-  bool verbose_output = false;
-};
-
-inline std::ostream& operator<< (std::ostream& str, const InitialPartitioningParameters& params) {
-  str << "-------------------------------------------------------------------------------"
-      << std::endl;
-  str << "Initial Partitioning Parameters:" << std::endl;
-  str << "  # IP trials:                        " << params.nruns << std::endl;
-  str << "  Mode:                               " << params.mode << std::endl;
-  str << "  Technique:                          " << params.technique << std::endl;
-  str << "  Algorithm:                          " << params.algo << std::endl;
-  if (params.technique == InitialPartitioningTechnique::multilevel) {
-    str << "IP Coarsening:                        " << std::endl;
-    str << params.coarsening;
-    str << "IP Local Search:                      " << std::endl;
-    str << params.local_search;
-  }
-  str << "-------------------------------------------------------------------------------"
-      << std::endl;
-  return str;
-}
-
-struct PartitioningParameters {
-  Mode mode = Mode::UNDEFINED;
-  Objective objective = Objective::UNDEFINED;
-  double epsilon = std::numeric_limits<double>::max();
-  PartitionID k = std::numeric_limits<PartitionID>::max();
-  PartitionID rb_lower_k = 0;
-  PartitionID rb_upper_k = 0;
-  int seed = 0;
-  uint32_t global_search_iterations = std::numeric_limits<uint32_t>::max();
-  int time_limit = 0;
-
-  mutable uint32_t current_v_cycle = 0;
-  std::vector<HypernodeWeight> perfect_balance_part_weights;
-  std::vector<HypernodeWeight> max_part_weights;
-  HyperedgeID hyperedge_size_threshold = std::numeric_limits<HypernodeID>::max();
-
-  bool verbose_output = false;
-  bool quiet_mode = false;
-  bool sp_process_output = false;
-  bool use_individual_part_weights = false;
-  bool vcycle_refinement_for_input_partition = false;
-  bool write_partition_file = true;
-
-  std::string graph_filename { };
-  std::string graph_partition_filename { };
-  std::string fixed_vertex_filename { };
-  std::string input_partition_filename { };
-};
-
-inline std::ostream& operator<< (std::ostream& str, const PartitioningParameters& params) {
-  str << "Partitioning Parameters:" << std::endl;
-  str << "  Hypergraph:                         " << params.graph_filename << std::endl;
-  str << "  Partition File:                     " << params.graph_partition_filename << std::endl;
-  if (!params.fixed_vertex_filename.empty()) {
-    str << "  Fixed Vertex File:                  " << params.fixed_vertex_filename << std::endl;
-  }
-  if (!params.input_partition_filename.empty()) {
-    str << "  Input Partition File:                  " << params.input_partition_filename << std::endl;
-  }
-  str << "  Mode:                               " << params.mode << std::endl;
-  str << "  Objective:                          " << params.objective << std::endl;
-  str << "  k:                                  " << params.k << std::endl;
-  str << "  epsilon:                            " << params.epsilon << std::endl;
-  str << "  seed:                               " << params.seed << std::endl;
-  str << "  # V-cycles:                         " << params.global_search_iterations << std::endl;
-  str << "  time limit:                         " << params.time_limit << "s" << std::endl;
-  str << "  hyperedge size threshold:           " << params.hyperedge_size_threshold << std::endl;
-  str << "  use individual block weights:       " << std::boolalpha
-      << params.use_individual_part_weights << std::endl;
-  if (params.use_individual_part_weights) {
-    for (PartitionID i = 0; i < params.k; ++i) {
-      str << "  L_opt" << i << ":                             " << params.perfect_balance_part_weights[i]
-          << std::endl;
-    }
-  } else {
-    str << "  L_opt" << ":                              " << params.perfect_balance_part_weights[0]
-        << std::endl;
-  }
-  if (params.use_individual_part_weights) {
-    for (PartitionID i = 0; i < params.k; ++i) {
-      str << "  L_max" << i << ":                             " << params.max_part_weights[i]
-          << std::endl;
-    }
-  } else {
-    str << "  L_max" << ":                              " << params.max_part_weights[0]
-        << std::endl;
-  }
-  return str;
-}
-struct EvolutionaryParameters {
-  size_t population_size;
-  float mutation_chance;
-  float edge_frequency_chance;
-  EvoReplaceStrategy replace_strategy;
-  mutable EvoCombineStrategy combine_strategy = EvoCombineStrategy::UNDEFINED;
-  mutable EvoMutateStrategy mutate_strategy = EvoMutateStrategy::UNDEFINED;
-  int diversify_interval;  // -1 disables diversification
-  double gamma;
-  size_t edge_frequency_amount;
-  bool dynamic_population_size;
-  float dynamic_population_amount_of_time;
-  bool random_combine_strategy;
-  mutable int iteration;
-  mutable Action action;
-  const std::vector<PartitionID>* parent1 = nullptr;
-  const std::vector<PartitionID>* parent2 = nullptr;
-  mutable std::vector<size_t> edge_frequency;
-  mutable std::vector<ClusterID> communities;
-  bool unlimited_coarsening_contraction;
-  bool random_vcycles;
-  bool parallel_partitioning_quick_start;
-};
-
-inline std::ostream& operator<< (std::ostream& str, const EvolutionaryParameters& params) {
-  str << "Evolutionary Parameters:              " << std::endl;
-  str << "  Population Size:                    " << params.population_size << std::endl;
-  str << "  Mutation Chance                     " << params.mutation_chance << std::endl;
-  str << "  Edge Frequency Chance               " << params.edge_frequency_chance << std::endl;
-  str << "  Replace Strategy                    " << params.replace_strategy << std::endl;
-  str << "  Combine Strategy                    " << params.combine_strategy << std::endl;
-  str << "  Mutation Strategy                   " << params.mutate_strategy << std::endl;
-  str << "  Diversification Interval            " << params.diversify_interval << std::endl;
-  str << "  Parallel Population Generation      " << params.parallel_partitioning_quick_start << std::endl;
   return str;
 }
